requirements.txt
gradio
openai
pyyaml
python-dotenv
